+++ lib/src/base_model.dart
@@ -30,19 +30,19 @@ abstract class BaseModelParser<T extends BaseModel> {
 }
 
 abstract class BaseModelApi<T extends BaseModel> {
-  final BaseModelParser<T> _parser;
-  final BaseApiClient _apiClient;
+  final BaseModelParser<T> parser;
+  final BaseApiClient apiClient;
   final bool rootList = false;
 
   abstract String url;
 
-  BaseModelApi(this._apiClient, this._parser);
+  BaseModelApi(this.apiClient, this.parser);
 
   Future<T> retrieve(dynamic id, {bool protected = false}) async {
-    final Response response = await get(id.toString(), protected: protected);
+    final Response response = await get('/${id.toString()}', protected: protected);
     final Map data = response.data as Map;
 
-    return _parser.fromJson(data);
+    return parser.fromJson(data);
   }
 
   Future<List<T>> list(
@@ -51,7 +51,7 @@ abstract class BaseModelApi<T extends BaseModel> {
       final Response response =
           await get('', params: params, protected: protected);
       final List results = response.data;
-      return results.map((e) => _parser.fromJson(e as Map)).toList();
+      return results.map((e) => parser.fromJson(e as Map)).toList();
     }
 
     final Pagination<T> pagination =
@@ -67,7 +67,7 @@ abstract class BaseModelApi<T extends BaseModel> {
         await get('', params: params, protected: protected);
     final dynamic data = response.data;
 
-    return Pagination.fromJson(data, _parser.fromJson);
+    return Pagination.fromJson(data, parser.fromJson);
   }
 
   Future<T> create(T source, {bool protected = true}) async {
@@ -75,7 +75,7 @@ abstract class BaseModelApi<T extends BaseModel> {
         await post('', body: source.toJson(), protected: protected);
     final Map data = response.data as Map;
 
-    return _parser.fromJson(data);
+    return parser.fromJson(data);
   }
 
   Future<T> update(T source, {bool protected = true}) async {
@@ -86,28 +86,66 @@ abstract class BaseModelApi<T extends BaseModel> {
     );
     final Map data = response.data as Map;
 
-    return _parser.fromJson(data);
+    return parser.fromJson(data);
   }
 
-  Future<void> detele(dynamic id, {dynamic body, bool protected = true}) {
+  Future<void> destroy(dynamic id, {dynamic body, bool protected = true}) {
     return delete('/$id', body: body, protected: protected);
   }
 
-  Future<Response> get(String action,
-      {Map<String, dynamic>? params, bool protected = false}) {
-    return _apiClient.get('$url$action/', params: params, protected: protected);
+  Future<Response> get(
+    String action, {
+    Map<String, dynamic>? params,
+    bool protected = false,
+    void Function(int, int)? onReceiveProgress,
+  }) {
+    return apiClient.get(
+      '$url$action/',
+      params: params,
+      protected: protected,
+      onReceiveProgress: onReceiveProgress,
+    );
   }
 
-  Future<Response> post(String action, {dynamic body, bool protected = true}) {
-    return _apiClient.post('$url$action/', body: body, protected: protected);
+  Future<Response> post(
+    String action, {
+    dynamic body,
+    bool protected = true,
+    void Function(int, int)? onSendProgress,
+    void Function(int, int)? onReceiveProgress,
+  }) {
+    return apiClient.post(
+      '$url$action/',
+      body: body,
+      protected: protected,
+    );
   }
 
-  Future<Response> put(String action, {dynamic body, bool protected = true}) {
-    return _apiClient.put('$url$action/', body: body, protected: protected);
+  Future<Response> put(
+    String action, {
+    dynamic body,
+    bool protected = true,
+    void Function(int, int)? onSendProgress,
+    void Function(int, int)? onReceiveProgress,
+  }) {
+    return apiClient.put(
+      '$url$action/',
+      body: body,
+      protected: protected,
+      onSendProgress: onSendProgress,
+      onReceiveProgress: onReceiveProgress,
+    );
   }
 
-  Future<Response> delete(String action,
-      {dynamic body, bool protected = true}) {
-    return _apiClient.delete('$url$action/', body: body, protected: protected);
+  Future<Response> delete(
+    String action, {
+    dynamic body,
+    bool protected = true,
+  }) {
+    return apiClient.delete(
+      '$url$action/',
+      body: body,
+      protected: protected,
+    );
   }
 }